requirements.txt
streamlit>=1.31,<2.0
google-generativeai>=0.8.0,<1.0
python-dotenv>=1.0.1,<2.0
requests>=2.31,<3.0
